requirements.txt
Django==5.2.8
cloudinary==1.44.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.0.1
Pillow==12.0.0
psycopg2-binary==2.9.11
black==25.9.0
flake8==7.3.0
Flake8-pyproject==1.2.3
isort==7.0.0
gunicorn==23.0.0
whitenoise==6.8.2
